+++ services/meeting/v1/model/model_record_info_req.go
@@ -13,10 +13,10 @@ type RecordInfoReq struct {
 	ConfUUID *string `json:"confUUID,omitempty"`
 
 	// 录制段落查询偏移量
-	SegmentOffset *int64 `json:"segmentOffset,omitempty"`
+	SegmentOffset *int32 `json:"segmentOffset,omitempty"`
 
 	// 录制段落查询数量
-	SegmentLimit *int64 `json:"segmentLimit,omitempty"`
+	SegmentLimit *int32 `json:"segmentLimit,omitempty"`
 }
 
 func (o RecordInfoReq) String() string {
